Drop backend cookie auth path; make session a single source of truth
The notifications backend only calls public endpoints, so remove the
speculative server-side auth from @lfc/api: createClient no longer takes
getSessionToken or sets a Cookie header. The client is now purely the API
surface; auth is entirely the mobile app's concern via credentials:'include'.

Consolidate the mobile session into one owner (~/lib/session): it holds the
token in SecureStore (persistence), the native cookie store (transport), and an
in-memory snapshot, and every mutation (setSession/clearSession/restoreSession)
updates all three together so they can't drift. Previously the token was
written in three places wired together by auth-context in a specific order.

- api.ts: drops the cookie glue, just binds the client to config + the 401
  handler and re-exports.
- session-store.ts: removed, folded into session.ts (which also owns the
  external-store subscribe/getSnapshot the UI reads).
- auth-context.tsx: now just orchestrates the flow (sign in/out, restore, 401)
  on top of the session module.

typecheck (all packages), mobile eslint and formatting pass.
