+++ apps/mobile/src/api.ts
@@ -1,38 +1,16 @@
 import { createClient } from '@lfc/api'
-import CookieManager from '@react-native-cookies/cookies'
 import { config } from '~/config'
 
 ///////////////////////////////////////////////////////////
-// Native API adapter
+// API client
 ///////////////////////////////////////////////////////////
 //
 // The transport-agnostic API layer lives in `@lfc/api` (shared with the
-// notifications backend). The only genuinely native concern is *session
-// transport*: on iOS the networking layer manages cookies through its own
-// store and silently drops a hand-set `Cookie` header, so the session has to
-// live in the native cookie store and go out via `credentials: 'include'`.
-//
-// This module owns that native glue and re-exports the client's methods, so the
-// rest of the app keeps importing everything from `~/api` as before.
-
-const API_URL = config.get('apiUrl')
-const COOKIE_ORIGIN = new URL(API_URL).origin
-
-/**
- * Set (or clear) the session cookie used to authenticate API requests. The auth
- * layer is the source of truth and keeps this in sync with persisted storage.
- */
-export async function setSessionToken(token: string | null) {
-  if (token) {
-    await CookieManager.set(COOKIE_ORIGIN, {
-      name: 'lfc-se',
-      value: token,
-      path: '/',
-    })
-  } else {
-    await CookieManager.clearAll()
-  }
-}
+// notifications backend). This module binds it to the app's config and re-
+// exports its methods, so the rest of the app keeps importing everything from
+// `~/api`. Session handling — persisting the token and putting it in the native
+// cookie store that `credentials: 'include'` reads — lives in `~/lib/session`,
+// the single source of truth.
 
 // Invoked whenever a request comes back 401, so the auth layer can drop an
 // expired session. Registered by the auth store; see `setUnauthorizedHandler`.
@@ -47,8 +25,8 @@ export function setUnauthorizedHandler(handler: (() => void) | null) {
 }
 
 const client = createClient({
-  baseUrl: API_URL,
-  // The session cookie lives in the native store (see `setSessionToken`);
+  baseUrl: config.get('apiUrl'),
+  // The session cookie lives in the native store (managed by `~/lib/session`);
   // `include` makes the native layer attach it on every request.
   credentials: 'include',
   onUnauthorized: () => onUnauthorized?.(),
